pyproject.toml
[tool.poetry]
name = "final"
version = "0.1.0"
description = ""
authors = ["Karl Sondeji <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
rich = "^11.2.0"
pyserde = "^0.7.0"
typer = "^0.4.0"
pandas = "^2.0.1"
numpy = "^1.24.0"
scipy = "^1.10.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
black = "^22.1.0"
pytest-cov = "^3.0.0"
pylama = "^8.3.7"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
